requirements.txt
streamlit
requests
pandas
numpy
matplotlib
seaborn
plotly
Pillow
streamlit-extras
nltk
wordcloud